requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
billiard==3.3.0.23
celery==3.1.23
contextlib2==0.5.4
cookies==2.2.1
coverage==4.0.2
cssselect==0.9.2
decorator==4.0.10
defusedxml==0.4.1
Django==1.8.14
django-admin-smoke-tests==0.2.1
django-anymail==0.4.2
django-countries==3.4.1
-e git+https://github.com/f213/django-suit-daterange-filter@a9d9c2e118f0006ca0fa842febbe0922fba75518#egg=django_date_range_filter
django-debug-toolbar==1.4
django-environ==0.4.0
django-mail-templated==2.6.2
django-markdown==0.8.4
django-money==0.8
django-nose==1.4.4
django-picklefield==0.3.2
django-redis==4.4.4
django-redis-cache==1.6.5
django-suit==0.2.21
django-timezone-field==1.3
django-user-agents==0.3.0
django-with-asserts==0.0.1
docutils==0.12
fake-factory==0.5.7
flake8==3.0.4
geoip2==2.4.0
gnureadline==6.3.3
icalendar==3.10
ipaddress==1.0.14
ipython==4.2.0
ipython-genutils==0.1.0
isort==4.2.5
kombu==3.0.35
lxml==3.6.0
Markdown==2.6.6
maxminddb==1.2.1
mccabe==0.5.0
mixer==5.5.7
nose==1.3.7
numpy==1.11.1
oauthlib==1.1.2
pexpect==4.1.0
pickleshare==0.7.2
Pillow==3.3.0
psycopg2==2.6.1
ptyprocess==0.5.1
py-moneyed==0.6.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycodestyle==2.0.0
pyflakes==1.2.3
PyJWT==1.4.0
python-dateutil==2.5.3
python-social-auth==0.2.19
python3-openid==3.0.10
pytz==2015.7
PyYAML==3.11
raven==5.27.1
redis==2.10.5
requests==2.10.0
requests-oauthlib==0.6.1
responses==0.5.1
rsa==3.4.2
selenium==2.53.5
simplegeneric==0.8.1
simplejson==3.8.2
six==1.10.0
sqlparse==0.1.19
timezonefinder==1.5.7
traitlets==4.2.1
ua-parser==0.7.1
uritemplate==0.6
user-agents==1.0.1
